Start a trainin game, from within package game
This is to circumvent the problem that game points to AI-player, and
AI-player back to game. Bit ugly, cause I would like that all AI-stuff
was inside its own package... Maybe a nicer solution will be build in
the future.

Also fix some small game breaking bugs, suchs as the AI coming up with
illegal moves.